feat: add coordination session client API
Introduce Coordination Service Session synchronous API. For now, it
supports operations that do not create subscriptions (though, the
DescribeSemaphore method still can be used to observe owners and waiters
of a semaphore).

The Session object is a high-level abstraction of the session concept,
which is introduced in the Coordination Service API. The client takes
over and controls the network, connection and server issues. It also
seamlessly integration with the Golang context concept using the Lease
object which context is alive until the corresponding semaphore is
considered acquired.

See also #53